Time: 3 ms (75.31%) | Memory: 69.8 MB (25.86%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.java
@@ -1,11 +1,11 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-        double sum=0;
+        long sum=0;
         for(int i=0;i<k;i++){
             sum+=nums[i];
 
         }
-        double maxsum=sum;
+        long maxsum=sum;
         for(int i=k;i<nums.length;i++){
             sum=sum-nums[i-k]+nums[i];
             maxsum=Math.max(maxsum,sum);
